Omar

+++ cuisinetocrop/app/_lib/mongo/models/user.js
@@ -3,7 +3,7 @@ const mongoose = require( 'mongoose')
 const userSchema = new mongoose.Schema ({
     userID: {type:String, required:true, unique:true},
     restaurantName:{type:String, required:true}, 
-    zipCode: {type:Int64, required:true}
+    zipCode: {type:Int32, required:true}
 
 });
 export const User = mongoose.models.User || mongoose.model("User", userSchema);
